Now using big decimal pretty much everywhere

+++ backend/src/main/kotlin/com/techso/techsport/model/response/BaseTeamResponse.kt
@@ -1,9 +1,10 @@
 package com.techso.techsport.model.response
 
+import java.math.BigDecimal
 import java.time.LocalDate
 
 abstract class BaseTeamResponse(
     val name: String,
     val members: List<String>,
-    val pointChanges: Map<LocalDate, Double>
+    val pointChanges: Map<LocalDate, BigDecimal>
 )
